+++ commands/sleep.go
@@ -23,65 +23,41 @@
 package commands
 
 import (
-	"bytes"
-	"encoding/base64"
 	"fmt"
-	"net/http"
+	"strings"
+	"time"
+
+	"github.com/spf13/cobra"
 
 	"github.com/mkloubert/go-package-manager/types"
 	"github.com/mkloubert/go-package-manager/utils"
-	"github.com/spf13/cobra"
 )
 
-func Init_Base64_Command(parentCmd *cobra.Command, app *types.AppContext) {
-	var dataURI bool
-
-	var base64Cmd = &cobra.Command{
-		Use:     "base64 [files]",
-		Aliases: []string{"b64"},
-		Short:   "Encode Base64",
-		Long:    `Encode data from STDIN to STDOUT as Base64 encoded data.`,
+func Init_Sleep_Command(parentCmd *cobra.Command, app *types.AppContext) {
+	var sleepCmd = &cobra.Command{
+		Use:     "sleep [duration]",
+		Aliases: []string{"wait"},
+		Short:   "Sleeps awhile",
+		Long:    `Runs the application by doing noting for a specific duration.`,
 		Run: func(cmd *cobra.Command, args []string) {
-			debugWrittenValue := func(written int64) {
-				if app.Verbose {
-					fmt.Println()
-				}
-				app.Debug(fmt.Sprintf("Bytes written: %v", written))
+			duration := ""
+			if len(args) > 0 {
+				duration = strings.TrimSpace(args[0])
+			}
+			if duration == "" {
+				duration = "1s"
 			}
 
-			if dataURI {
-				var buffer bytes.Buffer
+			d, err := time.ParseDuration(duration)
+			utils.CheckForError(err)
 
-				written, err := app.WriteAllInputsTo(&buffer, args...)
-				utils.CheckForError(err)
+			app.Debug(fmt.Sprintf("Waiting %s ...", d.String()))
 
-				data := buffer.Bytes()
-				defer buffer.Reset()
-
-				encoder := base64.NewEncoder(base64.StdEncoding, app.Out)
-				defer encoder.Close()
-
-				mimeType := http.DetectContentType(data)
-				base64Data := base64.StdEncoding.EncodeToString(data)
-
-				app.Write([]byte(fmt.Sprintf("data:%s;base64,%s", mimeType, base64Data)))
-
-				debugWrittenValue(written)
-			} else {
-				encoder := base64.NewEncoder(base64.StdEncoding, app.Out)
-				defer encoder.Close()
-
-				written, err := app.WriteAllInputsTo(encoder, args...)
-				utils.CheckForError(err)
-
-				debugWrittenValue(written)
-			}
+			time.Sleep(d)
 		},
 	}
 
-	base64Cmd.Flags().BoolVarP(&dataURI, "data-uri", "", false, "output as data URI")
-
 	parentCmd.AddCommand(
-		base64Cmd,
+		sleepCmd,
 	)
 }